read streaming data
Instead of sending each value as a string, the app sends it as a
binary block. The size of the block can vary depending on the time it
takes to calculate the individual values of the Mandelbrot set. The
app reads newly calculated data in a specified time period and sends
it as a group of binary data via MQTT. These data are received as a
group and each value is output separately. This eliminates the need to
convert value to character and vice versa. A larger amount of data is
sent at once, which has a positive effect on the speed.